io/ompio: add a new grouping option avoiding communication
the new grouping option simple+ performs all calculations used
for the aggregator selection as if the default file view would be used,
thus avoiding communication in file_set_view all together. This mode
is useful for applications that do not set a file view, but use
explicit offset operations on the default file view.

This is the 2.x equivalen of commit 76a8c67575f1f554a5482e4265bea106002ac332  on
master and the 3.x branch

Signed-off-by: Edgar Gabriel <egabriel@central.uh.edu>