common/ompio: adjust interface of decode_datatype()
extend the interface of the decode_datatype function to take a datarepresentation
argument. This is further in preparation to support the external32 representation.
Note, that this function is used both for 'decoding' the fileview ( in which case
we have to honor the datarepresentation provided by the user) as well as for memory
datatypes ( in which case we hardcode to "native").

Signed-off-by: Edgar Gabriel <egabriel@central.uh.edu>