fbtl/posix: add io_uring fixed buffer support
add support for io_uring fixed buffers into the fbtl posix
component. The fixed buffer interfaces seem to give some
nice performance benefits if available.

The intention is to use them only in select scenarios, e.g.
for the aggregators buffers in collective I/O or as the staging
buffers for individual I/O with GPU buffers or external32
representation.

Signed-off-by: Edgar Gabriel <Edgar.Gabriel@amd.com>